+++ packages/nornj-react/antd/date-picker/index.js
@@ -2,13 +2,37 @@ import { registerComponent } from 'nornj';
 import DatePicker from 'antd/lib/date-picker';
 
 registerComponent({
-  'ant-DatePicker': DatePicker,
-  'ant-MonthPicker': DatePicker.MonthPicker,
-  'ant-WeekPicker': DatePicker.WeekPicker,
+  'ant-DatePicker': {
+    component: DatePicker,
+    options: {
+      fieldDefaultRule: {
+        type: 'date'
+      }
+    }
+  },
+  'ant-MonthPicker': {
+    component: DatePicker.MonthPicker,
+    options: {
+      fieldDefaultRule: {
+        type: 'date'
+      }
+    }
+  },
+  'ant-WeekPicker': {
+    component: DatePicker.WeekPicker,
+    options: {
+      fieldDefaultRule: {
+        type: 'date'
+      }
+    }
+  },
   'ant-RangePicker': {
     component: DatePicker.RangePicker,
     options: {
-      needToJS: true
+      needToJS: true,
+      fieldDefaultRule: {
+        type: 'array'
+      }
     }
   }
 });