+++ packages/nornj-react/src/interface.ts
@@ -1,3 +1,50 @@
+import schema, { RuleItem } from 'async-validator';
+import { IObservableObject } from 'mobx';
+
+export interface MobxFieldDataProps extends RuleItem {
+  name: string;
+  value?: any;
+  trigger?: string;
+  rules?: RuleItem[];
+  [key: string]: any;
+}
+
+export interface MobxFieldDataInstance extends MobxFieldDataProps {
+  setDefaultRule?(rule: RuleItem): void;
+  validatorSchema?: schema;
+  reset?: Function;
+  validateStatus?: string;
+  help?: string;
+}
+
+export interface MobxFieldData {
+  (props: MobxFieldDataProps): JSX.Element;
+}
+
+export interface MobxFormDataProps {
+  observable?: boolean;
+}
+
+export interface MobxFormDataInstance {
+  readonly _njMobxFormData: boolean;
+  fieldDatas: Map<string, MobxFieldDataInstance & IObservableObject>;
+  validate(name?: string | string[]): Promise<any>;
+  error(name: string, help: string): void;
+  clear(name?: string | string[]): void;
+  reset(name?: string | string[]): void;
+  add(fieldData: MobxFieldDataProps | JSX.Element): void;
+  delete(name: string): void;
+  setValue(name: string | object, value?: any): void;
+  formData: MobxFormDataInstance;
+  [key: string]: any;
+}
+
+export interface MobxFormData {
+  (props: MobxFormDataProps): JSX.Element;
+}
+
+export type JSXElementWithMobxFormData = { formData?: MobxFormDataInstance & IObservableObject };
+
 /**
  * React bindings for NornJ template engine.
  */