+++ packages/nornj/CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+## [v5.1.1] 2020.02.05
+
+* 🐞 Fixed a range operator bug in expressions. [#card-32606466](https://github.com/joe-sky/nornj/projects/1#card-32606466)
+
+## [v5.0.7] 2019.12.31
+
+* 🔧 Set the default value of `ExtensionOption.useExpressionInProps` to `false`.
+
 ## [v5.0.1] 2019.12.04
 
 * 🏷️ Added a `ElementType` type to replace the `React.ElementType` for compatibility.